lib_string.h: replace the argument type to uint, fix the comment.

+++ include/lib_string.h
@@ -28,13 +28,13 @@ int strcmp_w(UTF16 a, UTF16 b);
 // if a = b, return 0
 // if a > b, return 1
 // if a < b, return -1
-int strncmp_a(ANSI a, ANSI b, int64 n);
+int strncmp_a(ANSI a, ANSI b, uint n);
 
 // strncmp_w is used to compare two UTF-16 strings with length.
 // if a = b, return 0
 // if a > b, return 1
 // if a < b, return -1
-int strncmp_w(UTF16 a, UTF16 b, int64 n);
+int strncmp_w(UTF16 a, UTF16 b, uint n);
 
 // stricmp_a is used to compare two ANSI strings, it is case-insensitive.
 // 
@@ -53,13 +53,13 @@ int stricmp_w(UTF16 a, UTF16 b);
 // if a = b, return 0
 // if a > b, return 1
 // if a < b, return -1
-int strnicmp_a(ANSI a, ANSI b, int64 n);
+int strnicmp_a(ANSI a, ANSI b, uint n);
 
 // strnicmp_w is used to compare two UTF-16 strings with length, it is case-insensitive.
 // if a = b, return 0
 // if a > b, return 1
 // if a < b, return -1
-int strnicmp_w(UTF16 a, UTF16 b, int64 n);
+int strnicmp_w(UTF16 a, UTF16 b, uint n);
 
 // strcpy_a is used to copy source ANSI string to destination.
 uint strcpy_a(ANSI dst, ANSI src);
@@ -67,10 +67,10 @@ uint strcpy_a(ANSI dst, ANSI src);
 // strcpy_w is used to copy source UTF-16 string to destination.
 uint strcpy_w(UTF16 dst, UTF16 src);
 
-// strcpy_a is used to copy source ANSI string to destination with length.
-uint strncpy_a(ANSI dst, ANSI src, int64 n);
+// strncpy_a is used to copy source ANSI string to destination with length.
+uint strncpy_a(ANSI dst, ANSI src, uint n);
 
-// strcpy_w is used to copy source UTF-16 string to destination with length.
-uint strncpy_w(UTF16 dst, UTF16 src, int64 n);
+// strncpy_w is used to copy source UTF-16 string to destination with length.
+uint strncpy_w(UTF16 dst, UTF16 src, uint n);
 
 #endif // LIB_STRING_H
